chore(gitlab enrichment): rebase
