+++ .gitignore
@@ -135,3 +135,7 @@ Thumbs.db
 
 # yarn
 yarn.lock
+
+# rabbitmq
+
+rabbitmq
